refactoring the way we get `courseEvaluation` by wrapping from SWR
